FML
Added print statements to debug Applyderiva3a. Currently applyderive3a works when called multiple times by itself but when combining it with applyderiva1 for example adds chords that have the wrong duration and aren't the right chord type. This seems to happen only after 4 or more iterations of calling both applyderivations in a loop.